backend/requirements.txt
python-dotenv==1.0.1
openai==1.12.0
fastapi==0.109.2
uvicorn==0.27.1
groq==0.18.0
python-multipart==0.0.9
sqlalchemy==2.0.27
aiohttp==3.9.3
requests==2.31.0
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.4
python-dateutil==2.8.2
